requirements.txt
# Hands-on AI - Python Dependencies
# AI and ML frameworks
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# Vector stores and embeddings
chromadb>=0.4.22
# pinecone-client==3.0.0  # Optional alternative
# qdrant-client==1.7.0    # Optional alternative

# OpenAI
openai>=1.35.14

# Anthropic Claude
anthropic==0.8.1

# Google Gemini
google-generativeai>=0.8.3

# Data processing
pandas==2.1.4
numpy==1.26.3

# NLP and text processing
tiktoken==0.5.2
sentence-transformers==2.2.2

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2

# File formats
pyyaml==6.0.1
toml==0.10.2
jsonschema==4.21.1

# Code analysis and security
bandit==1.7.6
safety>=3.0.1
semgrep==1.55.0

# Linting and formatting
ruff==0.1.14
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7

# CLI utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Environment and configuration
python-dotenv==1.0.0
pydantic>=2.8.2
pydantic-settings>=2.3.0

# Database drivers (optional, uncomment as needed)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.1          # MongoDB
# redis==5.0.1            # Redis

# API frameworks (optional)
# fastapi==0.109.0
# uvicorn==0.27.0

# Jupyter (for development and experimentation)
jupyter==1.0.0
ipython==8.20.0

# Git integration
gitpython==3.1.41

# Date and time
python-dateutil==2.8.2

# Cryptography
cryptography==42.0.0

# Diff and patch utilities
python-Levenshtein==0.23.0

# Progress bars and UI
alive-progress==3.1.5

# Documentation generation
mkdocs==1.5.3
mkdocs-material==9.5.4

# Code generation helpers
jinja2==3.1.3
